Organize libdhdt
