+++ carst/libdhdt.py
@@ -1,23 +1,22 @@
-# Class: TimeSeriesDEM(np.ndarray)
-# Func: Resample_Array(UtilRaster.SingleRaster, UtilRaster.SingleRaster)
-# used for dhdt
+# Class: PixelTimeSeries, DemPile
+# Func: timeit, resample_array, EVMD group, wlr_corefun, onclick_ipynb
 # by Whyjay Zheng, Jul 27 2016
 # last edit: Apr 17 2023
 
 import numpy as np
 from numpy.linalg import inv
 import os
-import sys
+# import sys
 try:
-    import gdal
-except:
-    from osgeo import gdal        # sometimes gdal is part of osgeo modules
+    from osgeo import gdal
+except ImportError:
+    import gdal    # this was used until GDAL 3.1
 from datetime import datetime
 from shapely.geometry import Polygon
-from scipy.interpolate import interp2d
-from scipy.interpolate import NearestNDInterpolator
-from scipy.interpolate import griddata
-import gc
+# from scipy.interpolate import interp2d
+# from scipy.interpolate import NearestNDInterpolator
+# from scipy.interpolate import griddata
+# import gc
 from carst import ConfParams
 from carst.libraster import SingleRaster
 import pickle
@@ -37,6 +36,7 @@ def timeit(func):
 
 def resample_array(source, reference, method='bilinear', destination=None):
     """
+    latest version. 
     resample the source raster using the extent and spacing provided by the reference raster. Two rasters must be in the same CRS.
     
     source: class UtilRaster.SingleRaster object
@@ -46,7 +46,7 @@ def resample_array(source, reference, method='bilinear', destination=None):
     returns: an numpy array, which you can use the methods in UtilRaster to trasform it into a raster.
 
     For better COG processing efficiency, we rasterio to check the if the source and the reference intersect each other.
-    For fasting resampling, we use GDAL and it C library.
+    For fasting resampling, we use GDAL and its C library.
     """
     
     s_ulx, s_uly, s_lrx, s_lry = source.get_extent()
@@ -64,93 +64,12 @@ def resample_array(source, reference, method='bilinear', destination=None):
     else:
         return np.full((reference.get_y_size(), reference.get_x_size()), reference.get_nodata())
 
-def Resample_Array2(orig_dem, resamp_ref_dem, resampleAlg='bilinear'):
-
-    """
-    resample orig_dem using the extent and spacing provided by resamp_ref_dem
-    orig_dem: class UtilRaster.SingleRaster object
-    resamp_ref_dem: class UtilRaster.SingleRaster object
-    returns: an numpy array, which you can use the methods in UtilRaster to trasform it into a raster
-
-    This one uses gdal.Warp API.
-    """
-
-    ds = gdal.Open(orig_dem.fpath)
-    ulx, uly, lrx, lry = resamp_ref_dem.get_extent()
-    opts = gdal.WarpOptions(outputBounds=(ulx, lry, lrx, uly), xRes=resamp_ref_dem.get_x_res(), yRes=resamp_ref_dem.get_y_res(), resampleAlg=resampleAlg)
-    out_ds = gdal.Warp('tmp.tif', ds, options=opts)
-    return out_ds.GetRasterBand(1).ReadAsArray()
-
-
-def Resample_Array(orig_dem, resamp_ref_dem, resamp_method='linear'):
-
-	"""
-	resample orig_dem using the extent and spacing provided by resamp_ref_dem
-	orig_dem: class UtilRaster.SingleRaster object
-	resamp_ref_dem: class UtilRaster.SingleRaster object
-	returns: an numpy array, which you can use the methods in UtilRaster to trasform it into a raster
-
-	Uses linear interpolation because it best represent flat ice surface.
-	-9999.0 (default nan in a Geotiff) is used to fill area outside the extent of orig_dem.
-
-	resamp_method: 'linear', 'cubic', 'quintic', 'nearest'
-
-	"""
-	o_ulx, o_uly, o_lrx, o_lry = orig_dem.get_extent()
-	o_ulx, o_xres, o_xskew, o_uly, o_yskew, o_yres = orig_dem.GetGeoTransform()
-	orig_dem_extent = Polygon([(o_ulx, o_uly), (o_lrx, o_uly), (o_lrx, o_lry), (o_ulx, o_lry)])
-	ulx, uly, lrx, lry = resamp_ref_dem.get_extent()
-	ulx, xres, xskew, uly, yskew, yres = resamp_ref_dem.GetGeoTransform()
-	resamp_ref_dem_extent = Polygon([(ulx, uly), (lrx, uly), (lrx, lry), (ulx, lry)])
-	if orig_dem_extent.intersects(resamp_ref_dem_extent):
-		x = np.linspace(o_ulx, o_lrx - o_xres, orig_dem.get_x_size())
-		y = np.linspace(o_uly, o_lry - o_yres, orig_dem.get_y_size())
-		z = orig_dem.ReadAsArray()
-		# z[z == orig_dem.get_nodata()] = np.nan    # this line probably needs improvement
-		if resamp_method == 'nearest':
-			print('resampling method = nearest')
-			xx, yy = np.meshgrid(x, y)
-			points = np.stack((np.reshape(xx, xx.size), np.reshape(yy, yy.size)), axis=-1)
-			values = np.reshape(z, z.size)
-			fun = NearestNDInterpolator(points, values)
-			xnew = np.linspace(ulx, lrx - xres, resamp_ref_dem.get_x_size())
-			ynew = np.linspace(uly, lry - yres, resamp_ref_dem.get_y_size())
-			xxnew, yynew = np.meshgrid(xnew, ynew)
-			znew = fun(xxnew, yynew)    # if the image is big, this may take a long time (much longer than linear approach)
-		elif resamp_method == 'linear':
-			nan_value = orig_dem.get_nodata()
-			print('resampling method = interp2d - ' + resamp_method)
-			fun = interp2d(x, y, z, kind=resamp_method, bounds_error=False, copy=False, fill_value=nan_value)
-			xnew = np.linspace(ulx, lrx - xres, resamp_ref_dem.get_x_size())
-			ynew = np.linspace(uly, lry - yres, resamp_ref_dem.get_y_size())
-			znew = np.flipud(fun(xnew, ynew))    # I don't know why, but it seems f(xnew, ynew) is upside down.
-			fun2 = interp2d(x, y, z != nan_value, kind=resamp_method, bounds_error=False, copy=False, fill_value=0)
-			zmask = np.flipud(fun2(xnew, ynew))
-			znew[zmask <= 0.999] = nan_value
-		else:
-			# This is deprecated...
-			print('resampling method = griddata - ' + resamp_method)
-			xx, yy = np.meshgrid(x, y)
-			xx = xx.ravel()
-			yy = yy.ravel()
-			zz = z.ravel()
-			realdata_pos = zz != orig_dem.get_nodata()
-			xx = xx[realdata_pos]
-			yy = yy[realdata_pos]
-			zz = zz[realdata_pos]
-			xnew = np.linspace(ulx, lrx - xres, resamp_ref_dem.get_x_size())
-			ynew = np.linspace(uly, lry - yres, resamp_ref_dem.get_y_size())
-			znew = griddata((xx, yy), zz, (xnew[None,:], ynew[:,None]), method='linear')
-		del z
-		gc.collect()
-		return znew
-	else:
-		return np.ones_like(resamp_ref_dem.ReadAsArray()) * -9999.0
 
     
 def EVMD_DBSCAN(x, y, eps=6, min_samples=4):
     """
     See EVMD & EVMD_idx.
+    Recommended to use this inreplace of EVMD and EVMD_idx.
     """
     # x, assuming the temporal axis (unit=days), is scaled by 365 for proper eps consideration.
     x2d = np.column_stack((x / 365, y))
@@ -269,80 +188,6 @@ def wlr_corefun(x, y, ye, evmd_threshold=6, detailed=False):
                 return slope, slope_err, resid, count
 
 
-		# ============ Using Cook's Distance ============
-		# cookd     = ri2 / (p_num * mse) * (h / (1 - h) ** 2)    # Cook's Distance
-		# goodpoint = cookd < 4 / x.size                          # Bollen, Kenneth A.; Jackman, Robert W. (1990) (see wikipedia)
-		# # xnew = x[goodpoint]
-		# # ynew = y[goodpoint]
-		# # yenew = ye[goodpoint]
-		# # wnew = [1 / k for k in yenew]
-		# if np.all(goodpoint):
-		# 	slope = p[0] * 365.25
-		# 	slope_err = np.sqrt(cov_m[0, 0]) * 365.25
-		# 	count = x.size
-		# 	if resid > 100000:
-		# 		print(x,y,ye,cookd,goodpoint)
-		# 	return slope, slope_err, resid, count
-		# else:
-		# 	xnew = x[goodpoint]
-		# 	ynew = y[goodpoint]
-		# 	yenew = ye[goodpoint]
-		# 	return wlr_corefun(xnew, ynew, yenew)
-
-	# =========== Old function ===========
-	# if xnew.size > 4:
-	# 	p, c            = np.polyfit(xnew, ynew, 1, w=wnew, cov=True)
-	# 	_, res, _, _, _ = np.polyfit(xnew, ynew, 1, w=wnew, full=True)
-	# 	residual = np.sum((np.polyval(p, xnew) - ynew) ** 2)
-	# 	cov_m = c * (len(w) - 4) / res
-	# 	error =  np.sqrt(cov_m[0, 0])
-	# 	slope = p[0] * 365.25
-	# 	slope_err = np.sqrt(cov_m[0, 0]) * 365.25
-	# else:
-	# 	error = -9999.0
-	# 	slope = -9999.0
-	# 	slope_err = -9999.0
-
-	# case = 0
-	# w = [1 / k for k in ye]
-	# if len(x) == 4:
-	# 	case = 1
-	# 	# This is to avoid dividing by zero when N = 4 and to give us a better error estimate
-	# 	x  = np.append(x,   x[-1])
-	# 	y  = np.append(y,   y[-1])
-	# 	ye = np.append(ye, ye[-1])
-	# 	w  = np.append(w, sys.float_info.epsilon)
-	# elif len(x) == 3:
-	# 	case = 2
-	# 	# This is to avoid negative Cd^2 when N = 3 and to give us a better error estimate
-	# 	x  = np.append(x,  [x[-1],  x[-1]])
-	# 	y  = np.append(y,  [y[-1],  y[-1]])
-	# 	ye = np.append(ye, [ye[-1], ye[-1]])
-	# 	w  = np.append(w,  [sys.float_info.epsilon, sys.float_info.epsilon])
-	# p, c            = np.polyfit(x, y, 1, w=w, cov=True)
-	# _, res, _, _, _ = np.polyfit(x, y, 1, w=w, full=True)
-	# # where c is the covariance matrix of p -> c[0, 0] is the variance estimate of the slope.
-	# # what we want is ({G_w}^T G_w)^{-1}, which is equal to c * (N - m - 2) / res
-	# cov_m = c * (len(w) - 4) / res
-	# slope = p[0] * 365.25
-	# slope_err = np.sqrt(cov_m[0, 0]) * 365.25
-	# # slope_error_arr[m, n] = np.sqrt(c[0, 0]) * 365.25
-	# # ./point_TS_ver2-2_linreg.py:^^^^^^^^: RuntimeWarning: invalid value encountered in sqrt
-	# # /data/whyj/Software/anaconda3/lib/python3.5/site-packages/numpy/lib/polynomial.py:606: RuntimeWarning: divide by zero encountered in true_divide
-	# # fac = resids / (len(x) - order - 2.0)
-	# if case == 0:
-	# 	residual = np.sum((np.polyval(p, x) - y) ** 2)
-	# elif case == 1:
-	# 	residual = np.sum((np.polyval(p, x[:-1]) - y[:-1]) ** 2)
-	# elif case == 2:
-	# 	residual = np.sum((np.polyval(p, x[:-2]) - y[:-2]) ** 2)
-	# return slope, slope_err, residual
-
-
-
-
-
-
 class PixelTimeSeries(object):
     """
     Store all measurements of a single pixel in the reference DEM.
@@ -398,7 +243,7 @@ class DemPile(object):
     is stored in the intermediate pickle file. 
     """
 
-    def __init__(self, picklepath=None, refgeo=None, refdate=None, dhdtprefix=None):
+    def __init__(self, picklepath=None, refgeo=None, refdate=None, dhdtprefix=None, evmd_threshold=6):
         self.picklepath = picklepath
         self.dhdtprefix = dhdtprefix
         self.ts = None
@@ -413,7 +258,7 @@ class DemPile(object):
         self.fitdata = {'slope': [], 'slope_err': [], 'residual': [], 'count': []}
         self.mosaic = {'value': [], 'date': [], 'uncertainty': []}
         self.maskparam = {'max_uncertainty': 9999, 'min_time_span': 0}
-        self.evmd_threshold = 6
+        self.evmd_threshold = evmd_threshold
 
     def AddDEM(self, dems):
         # ==== Add DEM object list ====
@@ -488,7 +333,7 @@ class DemPile(object):
             if self.dems[i].uncertainty <= self.maskparam['max_uncertainty']:
                 datedelta = self.dems[i].date - self.refdate
                 # znew = Resample_Array(self.dems[i], self.refgeo, resamp_method='linear')
-                znew = Resample_Array2(self.dems[i], self.refgeo)
+                znew = resample_array(self.dems[i], self.refgeo)
                 znew_mask = np.logical_and(znew > 0, self.refgeomask)
                 fill_idx = np.where(znew_mask)
                 for m,n in zip(fill_idx[0], fill_idx[1]):
@@ -680,137 +525,4 @@ def onclick_wrapper(data, axs, evmd_threshold):
 
 
 
-class TimeSeriesDEM(np.ndarray):
-
-	"""
-	DEPRECATED IN DHDT v1.0.
-
-	This class can include many DEMs (in UtilRaster.SingleRaster object) and then create a 3-D matrix.
-	each DEM is aligned along z-axis so that TimeSeriesDEM[m, n, :] would be the time series at pixel (m, n).
-	"""
-
-	def __new__(cls, dem=None, array=None, date=None, uncertainty=None):
-
-		""" 
-		You need to provide a UtilRaster.SingleRaster object, or provide array, date, and uncertainty separately.
-		TimeSeriesDEM is a n-m-k ndarray matrix, which n-m is the dem dimension, k is the count of dems.
-		You need to make sure all input dems have the same size (pixel number).
-		TimeSeriesDEM.date: a list of datetime object, which the length is k.
-		TimeSeriesDEM.uncertainty: a list of uncertainty for each DEM. The length is also k.
-
-		example:
-		tsdem = TimeSeriesDEM(dem=foo)  ---> Add single DEM.  the method "AddDEM" also does the trick after tsdem is created.
-		tsdem = TimeSeriesDEM(array=bar, date=bar2, uncertainty=bar3)   ---> Add single DEMs or multiple DEMs.
-		        Note that bar.shape[2] == len(bar2) == len(bar3)
-
-		Refered to
-		http://docs.scipy.org/doc/numpy/user/basics.subclassing.html 
-		http://stackoverflow.com/questions/27910013/how-can-a-class-that-inherits-from-a-numpy-array-change-its-own-values
-		"""
-
-		if dem is not None:
-			# retrieve band 1 array, and then replace NoDataValue by np.nan
-			dem_array = dem.ReadAsArray()
-			dem_array[dem_array == dem.get_nodata()] = np.nan
-			obj = np.asarray(dem_array).view(cls)
-			obj.date = [dem.date]
-			obj.uncertainty = [dem.uncertainty]
-		elif all([arg is not None for arg in [array, date, uncertainty]]):
-			obj = np.asarray(array).view(cls)
-			obj.date = date if type(date) is list else [date]
-			obj.uncertainty = uncertainty if type(uncertainty) is list else [uncertainty]
-		else:
-			raise ValueError('You need to either set "dem" or set "array, date and uncertainty".') 
-		obj.daydelta = None
-		obj.weight = None
-		return obj
-
-	def __array_finalize__(self, obj):
-
-		""" See TimeSeriesDEM.__new__ for comments """
-
-		if obj is None: return
-		self.date = getattr(obj, 'date', None)
-		self.uncertainty = getattr(obj, 'uncertainty', None)
-		self.daydelta = getattr(obj, 'daydelta', None)
-		self.weight = getattr(obj, 'weight', None)
-
-
-	def AddDEM(self, dem):
-
-		"""
-		Add a new DEM to the DEM time series.
-		dem is a UtilRaster.SingleRaster object.
-		"""
-
-		self.date.append(dem.date)
-		self.uncertainty.append(dem.uncertainty)
-		# Add the first band, and then replace NoDataValue by np.nan
-		dem_array = dem.ReadAsArray()
-		dem_array[dem_array == dem.get_nodata()] = np.nan
-		return TimeSeriesDEM(array=np.dstack([self, dem_array]), date=self.date, uncertainty=self.uncertainty)
-
-	def Date2DayDelta(self):
-
-		"""
-		Make self.daydelta from [self.date - min(self.date)]
-		"""
-
-		t = np.array(self.date) - min(self.date)
-		self.daydelta = np.array([i.days for i in t])
-
-	def SetWeight(self):
-
-		"""
-		Weight is set to 1/sigma^2
-		"""
-
-		self.weight = 1.0 / np.array(self.uncertainty) ** 2
-
-	def Polyfit(self, min_count=5, min_time_span=365, min_year=2000, max_year=2016):
-
-		"""
-		Note that x and w are all 1-d array like, with the same length of the third dimension of the reg_array.
-		w is the weight, which is often set to the inverse of data covariance matrix, C_d^-1
-		Here w must have the same length of x.
-		"""
-
-		reg_size = list(self.shape)
-		pixel_count = reg_size[0] * reg_size[1]
-
-		y = self.reshape(pixel_count, reg_size[2]).T
-		slope         = np.zeros(pixel_count)
-		slope_err     = np.zeros(pixel_count)
-		intercept     = np.zeros(pixel_count)
-		intercept_err = np.zeros(pixel_count)
-		for i in range(y.shape[1]):
-			if i % 10000 == 0:
-				print("processing " + str(i) + " pixels out of " + str(y.shape[1]) + " pixels")
-			px_y = y[:, i]
-			valid_idx = ~np.isnan(px_y)
-			# judge if a pixel is able to do regression using the given arguments
-			minlim_idx = np.array(self.date) > datetime(min_year, 1, 1)
-			maxlim_idx = np.array(self.date) < datetime(max_year, 1, 1)
-			valid_idx = valid_idx & minlim_idx & maxlim_idx
-			if sum(valid_idx) < min_count:
-				slope[i] = slope_err[i] = intercept[i] = intercept_err[i] = np.nan
-			elif max(self.daydelta[valid_idx]) - min(self.daydelta[valid_idx]) < min_time_span:
-				slope[i] = slope_err[i] = intercept[i] = intercept_err[i] = np.nan
-			# begin the polyfits
-			else:
-				px_y = px_y[valid_idx]
-				px_x = self.daydelta[valid_idx]
-				px_w = self.weight[valid_idx]
-				# This method minimizes sum(w * (y_i - y^hat_i) ^ 2)
-				#    Here we set w=np.sqrt(px_w) becuase np.polyfit minimizes sum( (w * (y_i - y^hat_i)) ^ 2) by default.
-				# Covariance is estimated from multivariate t-distribution.
-				# Comparing to the v0.1 version, this new method is slightly more conservative
-				#    because it considers the case of small d.o.f.
-				p, c = np.polyfit(px_x, px_y, 1, w=np.sqrt(px_w), cov=True)
-				slope[i]         = p[0]
-				slope_err[i]     = np.sqrt(c[0, 0])
-				intercept[i]     = p[1]
-				intercept_err[i] = np.sqrt(c[1, 1])
-
-		return slope.reshape(reg_size[:-1]), intercept.reshape(reg_size[:-1]), slope_err.reshape(reg_size[:-1]), intercept_err.reshape(reg_size[:-1])
